Updated views for videos
