+++ videos/views.py
@@ -1,0 +1,26 @@
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.views.generic import ListView, DetailView
+
+from .models import Video
+
+
+class VideoListView(LoginRequiredMixin, ListView):
+    """
+    Creates the video list page
+    """
+    model = Video
+    template_name = 'resources/resource_list.html'
+    paginate_by = 9
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context['resources'] = 'videos'
+        return context
+
+
+class VideoDetailView(LoginRequiredMixin, DetailView):
+    """
+    Creates the video detail pages
+    """
+    model = Video
+    template_name = 'resources/video.html'
